test `slide_out` and `squash`
