+++ git_machete/tests/functional/test_machete.py
@@ -411,303 +411,6 @@ class TestMachete:
         )
 
     @mock.patch('git_machete.utils.run_cmd', mock_run_cmd)  # to hide git outputs in tests
-    def test_slide_out(self) -> None:
-        (
-            self.repo_sandbox.new_branch("develop")
-            .commit("develop commit")
-            .push()
-            .new_branch("slide_root")
-            .commit("slide_root_1")
-            .push()
-            .check_out("slide_root")
-            .new_branch("child_a")
-            .commit("child_a_1")
-            .push()
-            .check_out("slide_root")
-            .new_branch("child_b")
-            .commit("child_b_1")
-            .push()
-            .check_out("child_b")
-            .new_branch("child_c")
-            .commit("child_c_1")
-            .push()
-            .new_branch("child_d")
-            .commit("child_d_1")
-            .push()
-        )
-
-        self.launch_command("discover", "-y", "--roots=develop")
-
-        self.assert_command(
-            ["status", "-l"],
-            """
-            develop
-            |
-            | slide_root_1
-            o-slide_root
-              |
-              | child_a_1
-              o-child_a
-              |
-              | child_b_1
-              o-child_b
-                |
-                | child_c_1
-                o-child_c
-                  |
-                  | child_d_1
-                  o-child_d *
-            """,
-        )
-
-        # Slide-out a single interior branch with one downstream. (child_c)
-        # This rebases the single downstream onto the new upstream. (child_b -> child_d)
-
-        self.launch_command("go", "up")
-        self.launch_command("slide-out", "-n")
-
-        self.assert_command(
-            ["status", "-l"],
-            """
-            develop
-            |
-            | slide_root_1
-            o-slide_root
-              |
-              | child_a_1
-              o-child_a
-              |
-              | child_b_1
-              o-child_b
-                |
-                | child_d_1
-                o-child_d * (diverged from origin)
-            """,
-        )
-
-        # Slide-out an interior branch with multiple downstreams. (slide_root)
-        # This rebases all the downstreams onto the new upstream. (develop -> [child_a, child_b])
-        self.launch_command("traverse", "-Wy")
-        self.launch_command("go", "up")
-        self.launch_command("go", "up")
-
-        self.assert_command(
-            ["status", "-l"],
-            """
-                develop
-                |
-                | slide_root_1
-                o-slide_root *
-                  |
-                  | child_a_1
-                  o-child_a
-                  |
-                  | child_b_1
-                  o-child_b
-                    |
-                    | child_d_1
-                    o-child_d
-                """,
-        )
-
-        self.launch_command("slide-out", "-n")
-
-        self.assert_command(
-            ["status", "-l"],
-            """
-            develop
-            |
-            | child_a_1
-            o-child_a (diverged from origin)
-            |
-            | child_b_1
-            o-child_b * (diverged from origin)
-              |
-              | child_d_1
-              x-child_d
-            """,
-        )
-
-        self.launch_command("traverse", "-Wy")
-        self.assert_command(
-            ["status", "-l"],
-            """
-            develop
-            |
-            | child_a_1
-            o-child_a
-            |
-            | child_b_1
-            o-child_b *
-              |
-              | child_d_1
-              o-child_d
-            """,
-        )
-
-        # Slide-out a terminal branch. (child_d)
-        # This just slices the branch off the tree.
-        self.launch_command("go", "down")
-        self.launch_command("slide-out", "-n")
-
-        self.assert_command(
-            ["status", "-l"],
-            """
-            develop
-            |
-            | child_a_1
-            o-child_a
-            |
-            | child_b_1
-            o-child_b *
-            """,
-        )
-
-    @mock.patch('git_machete.utils.run_cmd', mock_run_cmd)  # to hide git outputs in tests
-    def test_squash_merge(self) -> None:
-        (
-            self.repo_sandbox.new_branch("root")
-            .commit("root")
-            .push()
-            .new_branch("develop")
-            .commit("develop")
-            .push()
-            .new_branch("feature")
-            .commit("feature_1")
-            .commit("feature_2")
-            .push()
-            .new_branch("child")
-            .commit("child_1")
-            .commit("child_2")
-            .push()
-        )
-
-        self.launch_command("discover", "-y", "--roots=root")
-
-        self.assert_command(
-            ["status", "-l"],
-            """
-            root
-            |
-            | develop
-            o-develop
-              |
-              | feature_1
-              | feature_2
-              o-feature
-                |
-                | child_1
-                | child_2
-                o-child *
-            """,
-        )
-
-        # squash-merge feature onto develop
-        (
-            self.repo_sandbox.check_out("develop")
-            .execute("git merge --squash feature")
-            .execute("git commit -m squash_feature")
-            .check_out("child")
-        )
-
-        # in default mode, feature is detected as "m" (merged) into develop
-        self.assert_command(
-            ["status", "-l"],
-            """
-            root
-            |
-            | develop
-            | squash_feature
-            o-develop (ahead of origin)
-              |
-              m-feature
-                |
-                | child_1
-                | child_2
-                o-child *
-            """,
-        )
-
-        # but under --no-detect-squash-merges, feature is detected as "x" (behind) develop
-        self.assert_command(
-            ["status", "-l", "--no-detect-squash-merges"],
-            """
-            root
-            |
-            | develop
-            | squash_feature
-            o-develop (ahead of origin)
-              |
-              | feature_1
-              | feature_2
-              x-feature
-                |
-                | child_1
-                | child_2
-                o-child *
-            """,
-        )
-
-        # traverse then slides out the branch
-        self.launch_command("traverse", "-w", "-y")
-        self.assert_command(
-            ["status", "-l"],
-            """
-            root
-            |
-            | develop
-            | squash_feature
-            o-develop
-              |
-              | child_1
-              | child_2
-              o-child *
-            """,
-        )
-
-        # simulate an upstream squash-merge of the feature branch
-        (
-            self.repo_sandbox.check_out("develop")
-            .new_branch("upstream_squash")
-            .execute("git merge --squash child")
-            .execute("git commit -m squash_child")
-            .execute("git push origin upstream_squash:develop")
-            .check_out("child")
-            .execute("git branch -D upstream_squash")
-        )
-
-        # status before fetch will show develop as out of date
-        self.assert_command(
-            ["status", "-l"],
-            """
-            root
-            |
-            | develop
-            | squash_feature
-            o-develop (behind origin)
-              |
-              | child_1
-              | child_2
-              o-child *
-            """,
-        )
-
-        # fetch-traverse will fetch upstream squash, detect, and slide out the child branch
-        self.launch_command("traverse", "-W", "-y")
-
-        self.assert_command(
-            ["status", "-l"],
-            """
-            root
-            |
-            | develop
-            | squash_feature
-            | squash_child
-            o-develop *
-            """,
-        )
-
-    @mock.patch('git_machete.utils.run_cmd', mock_run_cmd)  # to hide git outputs in tests
     def test_go_up(self) -> None:
         """Verify behaviour of a 'git machete go up' command.
 
@@ -2186,125 +1889,3 @@ class TestMachete:
 
         with pytest.raises(subprocess.CalledProcessError):
             self.repo_sandbox.check_out("mars")
-
-    def test_squash_with_valid_fork_point(self) -> None:
-        (
-            self.repo_sandbox.new_branch('branch-0')
-                .commit("First commit.")
-                .commit("Second commit.")
-        )
-        fork_point = get_current_commit_hash()
-
-        (
-            self.repo_sandbox.commit("Third commit.")
-                .commit("Fourth commit.")
-        )
-
-        self.launch_command('squash', '-f', fork_point)
-
-        expected_branch_log = (
-            "Third commit.\n"
-            "Second commit.\n"
-            "First commit."
-        )
-
-        current_branch_log = popen('git log -3 --format=%s')
-        assert current_branch_log == \
-            expected_branch_log, \
-            ("Verify that `git machete squash -f <fork-point>` squashes commit"
-                " from one succeeding the fork-point until tip of the branch.")
-
-    def test_squash_with_invalid_fork_point(self) -> None:
-        (
-            self.repo_sandbox.new_branch('branch-0')
-                .commit()
-                .new_branch('branch-1a')
-                .commit()
-        )
-        fork_point_to_branch_1a = get_current_commit_hash()
-
-        (
-            self.repo_sandbox.check_out('branch-0')
-                .new_branch('branch-1b')
-                .commit()
-        )
-
-        with pytest.raises(SystemExit):
-            # First exception MacheteException is raised, followed by SystemExit.
-            self.launch_command('squash', '-f', fork_point_to_branch_1a)
-
-    @mock.patch('git_machete.utils.run_cmd', mock_run_cmd_and_forward_stdout)
-    def test_slide_out_with_valid_down_fork_point(self) -> None:
-        (
-            self.repo_sandbox.new_branch('branch-0')
-                .commit()
-                .new_branch('branch-1')
-                .commit()
-                .new_branch('branch-2')
-                .commit()
-                .new_branch('branch-3')
-                .commit()
-                .commit('Second commit on branch-3.')
-        )
-        hash_of_second_commit_on_branch_3 = get_current_commit_hash()
-        self.repo_sandbox.commit("Third commit on branch-3.")
-
-        self.launch_command('discover', '-y')
-        self.launch_command(
-            'slide-out', '-n', 'branch-1', 'branch-2', '-d',
-            hash_of_second_commit_on_branch_3)
-
-        expected_status_output = (
-            """
-            branch-0 (untracked)
-            |
-            | Third commit on branch-3.
-            o-branch-3 * (untracked)
-            """
-        )
-
-        self.assert_command(['status', '-l'], expected_status_output)
-
-    def test_slide_out_with_invalid_down_fork_point(self) -> None:
-        (
-            self.repo_sandbox.new_branch('branch-0')
-                .commit()
-                .new_branch('branch-1')
-                .commit()
-                .new_branch('branch-2')
-                .commit()
-                .new_branch('branch-3')
-                .commit()
-                .check_out('branch-2')
-                .commit('Commit that is not ancestor of branch-3.')
-        )
-        hash_of_commit_that_is_not_ancestor_of_branch_2 = get_current_commit_hash()
-
-        self.launch_command('discover', '-y')
-
-        with pytest.raises(SystemExit):
-            self.launch_command(
-                'slide-out', '-n', 'branch-1', 'branch-2', '-d',
-                hash_of_commit_that_is_not_ancestor_of_branch_2)
-
-    def test_slide_out_with_down_fork_point_and_multiple_children_of_last_branch(self) -> None:
-        (
-            self.repo_sandbox.new_branch('branch-0')
-                .commit()
-                .new_branch('branch-1')
-                .commit()
-                .new_branch('branch-2a')
-                .commit()
-                .check_out('branch-1')
-                .new_branch('branch-2b')
-                .commit()
-        )
-
-        hash_of_only_commit_on_branch_2b = get_current_commit_hash()
-
-        self.launch_command('discover', '-y')
-
-        with pytest.raises(SystemExit):
-            self.launch_command(
-                'slide-out', '-n', 'branch-1', '-d',
-                hash_of_only_commit_on_branch_2b)
